fix tcp flags for single,sequence,random

+++ app/pktgen-seq.h
@@ -42,9 +42,9 @@ typedef struct pkt_seq_s {
     };
     uint16_t ether_hdr_size; /**< Size of Ethernet header in packet for VLAN ID */
 
-    uint32_t tcp_seq;  /**< TCP sequence number */
-    uint32_t tcp_ack;  /**< TCP acknowledge number*/
-    uint8_t tcp_flags; /**< TCP flags value */
+    uint32_t tcp_seq;   /**< TCP sequence number */
+    uint32_t tcp_ack;   /**< TCP acknowledge number*/
+    uint16_t tcp_flags; /**< TCP flags value */
 
     uint32_t mpls_entry;   /**< MPLS entry if used */
     uint16_t qinq_outerid; /**< Outer VLAN ID if Q-in-Q */
